Added Exception using @ExceptionHandler

+++ src/main/java/org/fintexel/supplier/repository/VendorRepo.java
@@ -3,6 +3,6 @@ package org.fintexel.supplier.repository;
 import org.fintexel.supplier.entity.Vendor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VendorRepo extends JpaRepository<Vendor,Integer>{
+public interface VendorRepo extends JpaRepository<Vendor,Long>{
 
 }
